Floodsms/requirements.txt
africastalking==1.2.9
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
schema==0.7.7
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
